Initital Commit
In this initial commit, I created the necessary basic files and structures for building a link aggregator application. The purpose of this commit is to mark the beginning of the application development and establish a solid foundation for the continuation of the project. With the basic structures and functionalities already implemented, it will be possible to advance the application development and add new features in an organized and efficient way.