Addition of the "assets" folder
In this commit, the "assets" folder was added containing the necessary files for the proper functioning of the application. Additionally, the "images" folder was also added inside the "assets" folder, containing the necessary images for the application to function correctly.
This commit is important to ensure that the application has all the necessary files for its proper functioning and to facilitate its development and maintenance.